Increased size of filename buffer for entity transition file in missionpack DLL from MAX_QPATH to MAX_OSPATH. Changed target_changelevel clear inventory check to use MAPTYPE_CUSTOM check instead of != MAPTYPE_ID. Misc improvements to func_door_secret2. Partial fix to func_door_secret as a movewith child.

+++ missionpack/p_text.c
@@ -304,7 +304,7 @@ void Do_Text_Display (edict_t *activator, int flags, char *message)
 			Com_sprintf(pakfile, sizeof(pakfile), "%s\\pak%d.pak", filename, i);
 			if (NULL != (f = fopen(pakfile, "rb")))
 			{
-				num = fread(&pakheader,1,sizeof(pak_header_t),f);
+				num = (int)fread(&pakheader, 1, sizeof(pak_header_t), f);
 				if (num >= sizeof(pak_header_t))
 				{
 					if ( pakheader.id[0] == 'P' &&
